% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_global_smooth}
\alias{draw_global_smooth}
\title{Draw global smooth of AFQ tract.}
\usage{
draw_global_smooth(plot_obj, attr_num, tract, plot_title, out_dir)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains
global smooth (int)}

\item{tract}{AFQ tract name (str)}

\item{plot_title}{Title of plot (str)}

\item{out_dir}{Path to output location}
}
\description{
Draw global smooth resulting from GS model. Use plot(sm(obj, int))
to extract values, calculate confidence intervals, and ggplot
on a scaled X-axis.
}
\details{
Writes <out_dir>/Plot_GAM_<tract>_Global.png
}
