% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_group_intx}
\alias{draw_group_intx}
\title{Draw behavior-nodeID interactions by group.}
\usage{
draw_group_intx(df, gam_obj, tract, y_var, y_name, plot_title, out_dir)
}
\arguments{
\item{df}{Tract dataframe supplied to GAM}

\item{gam_obj}{Returned object from GAM/BAM tool}

\item{tract}{AFQ tract name (str)}

\item{y_var}{Column name of behavior of interest (str)}

\item{y_name}{Y-axis title (str)}

\item{plot_title}{Title for plot (str)}

\item{out_dir}{Path to output location (str)}
}
\description{
Deprecated. Plot factorial 3D interaction between nodeID, dti_fa, and
behavior as a function of diagnosis group.
}
\details{
Writes <out_dir>/Plot_GAM_<tract>_Group-Intx_<y_var>.png
}
