% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_group_intx_diff}
\alias{draw_group_intx_diff}
\title{Draw 3D interaction difference smooth for experimental group.}
\usage{
draw_group_intx_diff(
  plot_obj,
  attr_num,
  tract,
  y_var,
  y_name,
  plot_title,
  out_dir
)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains
reference group difference interaction smooth (int)}

\item{tract}{AFQ tract name (str)}

\item{y_var}{Behavior of interest, used for Y-axis (str)}

\item{y_name}{Y-axis title (str)}

\item{plot_title}{Title of plot (str)}

\item{out_dir}{Path to output location}
}
\description{
Using the output of an ordered-factor group interaction
model, draw how group B differs in their nodeID-FA-continuous
interaction from the reference group (group A).
}
\details{
Writes <out_dir>/Plot_GAM_<tract>_Group-Intx-Diff_<y_var>.png
}
