% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_group_smooth}
\alias{draw_group_smooth}
\title{Draw group smooths.}
\usage{
draw_group_smooth(plot_obj, attr_num, tract, out_dir)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains
group smooths (int)}

\item{tract}{AFQ tract name (str)}

\item{out_dir}{Path to output location}
}
\description{
Draw group smooths resulting from a GS model.
}
\details{
Writes <out_dir>/Plot_GAM_<tract>_GS.png
}
