% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_group_smooth_diff}
\alias{draw_group_smooth_diff}
\title{Draw difference of group smooths of AFQ tract.}
\usage{
draw_group_smooth_diff(plot_obj, attr_num, tract, plot_title, out_dir)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains group
difference smooth (int)}

\item{tract}{AFQ tract name (str)}

\item{plot_title}{Title of plot (str)}

\item{out_dir}{Path to output location}
}
\description{
Plot an A-B difference smooth from an GAM using an ordered factor for group,
identify nodes which sig differ from 0, draw polygons to ID.
}
\details{
Writes <out_dir>/Plot_GAM_<tract>_GS-Diff.png
}
