% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_intx}
\alias{draw_intx}
\title{Draw interaction smooths of AFQ tract node, FA, and behavioral covariate.}
\usage{
draw_intx(plot_obj, attr_num, tract, y_name, plot_title, out_file)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains interaction
smooth (int)}

\item{tract}{AFQ tract name (str)}

\item{y_name}{Y-axis title (str)}

\item{plot_title}{Title of plot (str)}

\item{out_file}{Path to output location, file name}
}
\description{
Plot using the gamViz plottable object and corresponding attribute. Draws
tract node as the X-axis.
}
\details{
Writes <out_file>.png
}
