% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_intx_diff}
\alias{draw_intx_diff}
\title{Draw 3D interaction difference smooth for experimental group.}
\usage{
draw_intx_diff(plot_obj, attr_num, tract, y_name, plot_title, out_file)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains
reference group difference interaction smooth (int)}

\item{tract}{AFQ tract name (str)}

\item{y_name}{Y-axis title (str)}

\item{plot_title}{Title of plot (str)}

\item{out_file}{Path to output location, file name}
}
\description{
Using the output of an ordered-factor group interaction
model, draw how group B differs in their nodeID-FA-continuous
interaction from the reference group (group A). The plot is inverted
to illustrate B-A.
}
\details{
Writes <out_file>.png
}
