% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_smooth_intx}
\alias{draw_smooth_intx}
\title{Draw behavior-nodeID interactions.}
\usage{
draw_smooth_intx(plot_obj, attr_num, tract, y_var, y_name, plot_title, out_dir)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains interaction
smooth (int)}

\item{tract}{AFQ tract name (str)}

\item{y_var}{Behavior of interest, used for Y-axis (str)}

\item{y_name}{Y-axis title (str)}

\item{plot_title}{Title of plot (str)}

\item{out_dir}{Path to output location}
}
\description{
Plot the interaction of tract node, FA, and behavior covariate. Draws
tract node as the X-axis.
}
\details{
Writes <out_dir>/Plot_GAM_<tract>_Intx_<x_var>.png
}
