% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gams.R
\name{gam_G}
\alias{gam_G}
\title{Conduct basic GAM analysis, G method.}
\usage{
gam_G(df_tract, dist_fam)
}
\arguments{
\item{df_tract}{Long-formatted dataframe for AFQ tract, containing
columns of dti_fa, sex, subjectID, and nodeID.}

\item{dist_fam}{Desired family/distribution to use with data (gamma, beta,
or gaus; str).}
}
\value{
GAM object
}
\description{
Model dti_fa for a tract, where sex has a
separate intercept, subject is random effect,
and using cubic spline for nodeID. K=50 determined
by testing.
}
