% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gams.R
\name{gam_Gcov}
\alias{gam_Gcov}
\title{Conduct basic GAM analysis, G method with covariate.}
\usage{
gam_Gcov(df_tract, dist_fam, cov)
}
\arguments{
\item{df_tract}{Long-formatted dataframe for AFQ tract, containing
columns of dti_fa, sex, subjectID, and nodeID.}

\item{dist_fam}{Desired family/distribution to use with data (gamma, beta,
or gaus; str).}

\item{cov}{Covariate column name (str).}
}
\value{
GAM object
}
\description{
Similar to gam_model, but include covariate and control for
interaction of sex with covariate.
}
