% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gams.R
\name{gam_GS}
\alias{gam_GS}
\title{Model data with global + group smooths, GS method.}
\usage{
gam_GS(df_tract, dist_fam, col_group)
}
\arguments{
\item{df_tract}{Long-formatted dataframe for AFQ tract, containing
columns of dti_fa, sex, subjectID, and nodeID.}

\item{dist_fam}{Desired family/distribution to use with data (gamma, beta,
or gaus; str).}

\item{col_group}{Column name of factored grouping value (str).}
}
\value{
GAM object
}
\description{
Similar to gam_model, but each group factor has own smooth.
}
