% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gams.R
\name{gam_GSOF}
\alias{gam_GSOF}
\title{Model data with global + group smooths (ordered factors).}
\usage{
gam_GSOF(df, dist_fam, col_group)
}
\arguments{
\item{df}{Long-formatted dataframe for AFQ tract, containing
columns of dti_fa, sex, subjectID, and nodeID.}

\item{dist_fam}{Desired family/distribution to use with data (gamma, beta,
or gaus; str).}

\item{col_group}{Column name of ordered factored grouping value (str).}
}
\value{
GAM object
}
\description{
Similar to gam_GS_model, but use ordered factors so a difference
smooth can be calculated.
}
