% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gams.R
\name{gam_intxOF_model}
\alias{gam_intxOF_model}
\title{Model data with interaction by group (ordered factor).}
\usage{
gam_intxOF_model(df_tract, dist_fam, col_group, cont_var)
}
\arguments{
\item{df_tract}{Long-formatted dataframe for AFQ tract, containing
columns of dti_fa, sex, subjectID, and nodeID.}

\item{dist_fam}{Desired family/distribution to use with data (gamma, beta,
or gaus; str).}

\item{col_group}{Column name of ordered factored grouping value (str).}

\item{cont_var}{Continuous variable, which will interact with
predicted (str).}
}
\value{
GAM object
}
\description{
Model tract FA values, allow for interaction of nodeID, group, and
continuous variable (LGI, PPI, etc). Use ordered factor so interaction
of <cont_var> and nodeID difference from reference group can be
calculated.
}
