% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gams.R
\name{switch_family}
\alias{switch_family}
\title{Switch for determining desired family.}
\usage{
switch_family(dist_fam)
}
\arguments{
\item{dist_fam}{Desired family/distribution to use with data (gamma, beta,
or gaus; str).}
}
\value{
Family option for GAM/BAM function
}
\description{
Switch for determining desired family.
}
