% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{switch_names}
\alias{switch_names}
\title{Switch for decoding AFQ tract and behavior names.}
\usage{
switch_names(name)
}
\arguments{
\item{name}{AFQ tract or behavior name (str)}
}
\value{
Name, in long form (str)
}
\description{
Switch for decoding AFQ tract and behavior names.
}
