% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_group_intx_ref}
\alias{draw_group_intx_ref}
\title{Draw group interaction reference (control) 3D smooth.}
\usage{
draw_group_intx_ref(plot_obj, attr_num, tract, y_var, out_dir)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains
reference group interaction smooth of reference group (int)}

\item{tract}{AFQ tract name (str)}

\item{y_var}{Behavior of interest, used for Y-axis (str)}

\item{out_dir}{Path to output location}
}
\description{
Using the output of an ordered-factor group interaction
model, draw how the reference group A interacts with continuous
metric, nodeID, and predicted FA (s(x)).
}
\details{
Writes <out_dir>/Plot_GAM_<tract>_Group-Intx-Ref_<y_var>.png
}
