% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gams.R
\name{draw_smooth_intx}
\alias{draw_smooth_intx}
\title{Draw behavior-nodeID interactions.}
\usage{
draw_smooth_intx(plot_obj, attr_num, tract, x_var, x_name, plot_title, out_dir)
}
\arguments{
\item{plot_obj}{Plotable object returned by getViz (object)}

\item{attr_num}{List/attribute number of plot_obj that contains interaction
smooth (int)}

\item{tract}{AFQ tract name (str)}

\item{x_var}{Behavior of interest, used for X-axis (str)}

\item{x_name}{X-axis title (str)}

\item{plot_title}{Title of plot (str)}

\item{out_dir}{Path to output location}
}
\description{
Draw behavior-nodeID interactions.
}
\details{
Writes <out_dir>/Plot_GAM_<tract>_Intx_<x_var>.png
}
